Add files via upload
This paper provides a theoretical introduction to game theory on networks. It presents the theoretical context and defines games on networks. Starting from such theoretical framework, the paper proposes several examples that illustrate how the
structure of Nash equilibria on network changes with respect to traditional bilateral games. We show that there are more equilibria on networks than in two-player games and an equilibrium on networks does not necessarily requires all connected individuals to play mutual best responses, a necessary condition for any bilateral Nash equilibrium. Last, using few examples, we propose a rationale for an equilibrium refinement concept based on an idea of strategic dependence, in which the network structure plays a key role.